+++ src/pages/MyPage/UserMyPage.jsx
@@ -1,35 +1,23 @@
 import React from 'react';
-import '../../global.css';
-import { useEffect, useRef } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { useToast } from '../../context/ToastContext';
 import { MOCK_DELIVERY } from '../../mocks/myPage.mock';
 
 import styles from './UserMyPage.module.css';
 import ProfileCard from '../../components/MyPage/ProfileCard';
 import DeliveryStatusCard from '../../components/MyPage/DeliveryStatusCard';
 import MyShoppingLinks from '../../components/MyPage/MyShoppingLinks';
+
+import useRequireAuth from '../../hooks/useRequireAuth';
 import useAuthStore from '../../store/useAuthStore';
 
 export default function UserMyPage() {
-  const showToast = useToast();
-  const navigate = useNavigate();
-  const hasShownRef = useRef(false);
-
   const user = useAuthStore((s) => s.user);
-  const isLoggedIn = useAuthStore((s) => s.isLoggedIn);
+  const isAllowed = useRequireAuth({
+    redirectTo: '/login',
+    message: '로그인이 필요합니다.',
+    types: 'error',
+  });
 
-  useEffect(() => {
-    if (!isLoggedIn) {
-      if (!hasShownRef.current) {
-        showToast('로그인이 필요합니다.', 'error');
-        hasShownRef.current = true;
-        navigate('/login');
-      }
-    }
-  }, [isLoggedIn, user, showToast, navigate]);
-
-  if (!isLoggedIn || !user) return null;
+  if (!isAllowed) return null;
 
   return (
     <div className={styles.page}>
@@ -40,7 +28,7 @@ export default function UserMyPage() {
         </section>
 
         <section className={styles.section}>
-          <h2 className={styles.sectionTitle}>배송 현황</h2>
+          <h2 className={styles.secutionTitle}>배송 현황</h2>
           <DeliveryStatusCard delivery={MOCK_DELIVERY} />
         </section>
